requirements.txt
streamlit
python-docx
openai
requests
pdfminer.six
docx2txt
weasyprint
PyPDF2
pdf2docx
spire.doc
instructor
pytesseract
pdf2image
htmldocx
anthropic